requirements.txt
langchain
python-dotenv
langchain-google-genai
langchain-community
arxiv
langchain-huggingface
text-generation
transformers
google-search-results
numexpr
langchainhub
sentencepiece
jinja2
bitsandbytes
accelerate
fastapi
langserve
unicorn
sse_starlette